requirements.txt
Django>=3.2.8,<3.3
uWSGI>=2.0.19.1,<2.1
djangorestframework>=3.14.0,<=3.15.0
